requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
awscli==1.29.21
blinker==1.6.2
blis==0.7.11
boto3==1.28.21
botocore==1.31.21
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.0
docutils==0.16
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.298
langcodes==3.3.0
langsmith==0.0.39
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.6
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pathy==0.10.2
Pillow==9.5.0
pip==22.3.1
preshed==3.0.9
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
pypdf==3.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
rsa==4.7.2
s3transfer==0.6.2
safetensors==0.3.3
setuptools==65.5.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
srsly==2.4.8
streamlit==1.27.0
sympy==1.12
tenacity==8.2.3
thinc==8.1.12
tiktoken==0.5.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.22.0
wasabi==1.1.2
wheel==0.38.4
yarl==1.9.2
zipp==3.17.0
